+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Most recent change on the bottom.
 - `include_file_as_baseline_config` for simple modifications of existing configs
 - `nequip-deploy --using-dataset` to support data-dependent deployment steps
 - Support for Gaussian Mixture Model uncertainty quantification (https://doi.org/10.1063/5.0136574)
+- `nequip-deploy build --checkpoint` to deploy specific checkpoints easily
 
 ### Changed
 - Always require explicit `seed`